requirements.txt
colorama
futures
pysocks
beautifulsoup4